+++ src/game/lara/lara_cheat.h
@@ -1,3 +1,4 @@
 #pragma once
 
+void __cdecl Lara_Cheat_EndLevel(void);
 void __cdecl Lara_Cheat_GetStuff(void);