+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## [Unreleased](https://github.com/LostArtefacts/TR2X/compare/stable...develop) - ××××-××-××
+- added new console commands:
+    - `/endlevel`
 - added an option to fix M16 accuracy while running (#45)
 - fixed secret rewards not displaying shotgun ammo (#159)
 - fixed sounds not playing under certain circumstances (#113, regression from 0.2)